Move fault logging to better location
